phase2-ai-resume-analyzer/requirements.txt
fastapi 
uvicorn 
python-multipart 
python-dotenv 
google-genai
pydantic
pypdf 
python-docx